Tests/LibWeb/Text/expected/input-value.txt

text: "pass"
hidden: "pass"
button: "pass"
checkbox: "pass"
file: "" (threw exception)
